style_transfer/requirements.txt
Pillow
torch
torchvision
flask
requests